Refactor project into multiple packages
