+++ Dockerfile
@@ -1,6 +1,6 @@
 # syntax=docker/dockerfile:1
 
-FROM golang:1.20
+FROM golang:1.24
 
 ARG CLIENT_SECRET
 ARG CLIENT_ID
@@ -8,11 +8,11 @@ ARG REDIRECT_URL
 ARG AUTH_URL
 ARG ACCREF_TOKEN_URL
 
-ENV CLIENT_SECRET $CLIENT_SECRET
-ENV CLIENT_ID $CLIENT_ID
-ENV REDIRECT_URL $REDIRECT_URL
-ENV AUTH_URL $AUTH_URL
-ENV ACCREF_TOKEN_URL $ACCREF_TOKEN_URL
+ENV CLIENT_SECRET=$CLIENT_SECRET
+ENV CLIENT_ID=$CLIENT_ID
+ENV REDIRECT_URL=$REDIRECT_URL
+ENV AUTH_URL=$AUTH_URL
+ENV ACCREF_TOKEN_URL=$ACCREF_TOKEN_URL
 
 # Set destination for COPY
 WORKDIR /app
