+++ fitbit/constants.go
@@ -1,0 +1,11 @@
+package fitbit
+
+const GET_ACTIVITY_LOG_LIST_PATH = "1/user/%s/activities/list.json"
+const FITBIT_BASE_URL = "https://api.fitbit.com"
+const GetProfileUrl = "https://api.fitbit.com/1/user/%s/profile.json"
+const GET_PROFILE_PATH = "1/user/%s/profile.json"
+const GetLifetimeStatsUrl = "https://api.fitbit.com/1/user/%s/activities.json"
+const GET_LIFETIME_STATS_PATH = "1/user/%s/activities.json"
+const GetBadgesUrl = "https://api.fitbit.com/1/user/%s/badges.json"
+const GET_BADGES_PATH = "1/user/%s/badges.json"
+const GetActivityLogListUrl = "https://api.fitbit.com/1/user/%s/activities/list.json"
