+++ fitbit/get_return_struct.go
@@ -1,0 +1,6 @@
+package fitbit
+
+type GetReturnStruct interface {
+	ActivityLogList | Profile | LifetimeStats | BadgeList
+}
+
