--- get_return_struct.go
@@ -1,6 +1,0 @@
-package main
-
-
-type GetReturnStruct interface {
-	ActivityLogList | Profile | LifetimeStats | BadgeList
-}
